src/requirements.txt
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0

pymongo==4.6.0
Flask-PyMongo==2.3.0
mongoengine==0.27.0

black==22.6.0
isort==5.10.1

toml==0.10.2
python-dateutil==2.8.2
python-dotenv==0.20.0
bcrypt==4.0.1
Werkzeug>=2.0.0,<3.0.0

gunicorn==20.1.0
